:recycle: KMP の AndroidApplicationPlugin に変更
