+++ app/android/src/androidMain/AndroidManifest.xml
