+++ app/android/src/androidMain/res/mipmap-hdpi/ic_launcher_foreground.webp
