+++ app/android/src/androidMain/res/mipmap-hdpi/ic_launcher_round.webp
