+++ app/android/src/androidMain/res/mipmap-mdpi/ic_launcher_foreground.webp
