+++ app/android/src/androidMain/res/mipmap-xhdpi/ic_launcher.webp
