+++ app/android/src/androidMain/res/mipmap-xxhdpi/ic_launcher_foreground.webp
