+++ app/android/src/androidMain/res/mipmap-xxxhdpi/ic_launcher_foreground.webp
