+++ app/android/src/androidMain/res/values/colors.xml
