+++ app/android/src/androidMain/res/values/ic_launcher_background.xml
