--- build-logic/src/main/kotlin/club/nito/primitive/AndroidApplicationPlugin.kt
@@ -1,20 +1,0 @@
-package club.nito.primitive
-
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-
-@Suppress("unused")
-class AndroidApplicationPlugin : Plugin<Project> {
-    override fun apply(target: Project) {
-        with(target) {
-            with(pluginManager) {
-                apply("com.android.application")
-            }
-
-            androidApplication {
-                setupAndroid()
-            }
-        }
-    }
-}
-
